requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.7.0
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
distro==1.9.0
Django==5.1.4
django-rest-framework==0.1.0
djangorestframework==3.15.2
frozenlist==1.5.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.11
langchain-core==0.3.24
langchain-text-splitters==0.3.2
langsmith==0.2.2
multidict==6.1.0
numpy==2.2.0
openai==1.57.2
orjson==3.10.12
packaging==24.2
propcache==0.2.1
pydantic==2.10.3
pydantic_core==2.27.1
pypdf==5.1.0
python-decouple==3.8
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlparse==0.5.3
tenacity==9.0.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.18.3
